Move broker into file
