+++ backend/app/broker.py
@@ -1,0 +1,12 @@
+import paho.mqtt.client as mqtt
+import paho.mqtt.publish as publish
+
+def on_connect(client, userdata, flags, reason_code, properties):
+    print(f"Connected with result code {reason_code}")
+    # Subscribing in on_connect() means that if we lose the connection and
+    # reconnect then subscriptions will be renewed.
+    client.subscribe("$SYS/#")
+
+mqttc = mqtt.Client(mqtt.CallbackAPIVersion.VERSION2)
+mqttc.on_connect = on_connect
+mqttc.connect("mqtt-broker", 1883, 60)
